% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutilities.r
\name{approxEqual}
\alias{approxEqual}
\title{checks whether two numbers are qpproximately the same to some absolute error}
\usage{
approxEqual(x, y, absError = 1e-08)
}
\arguments{
\item{x}{a number}

\item{y}{a number}

\item{absError}{the absolute error defaults to 1E-8}
}
\value{
whether the two numbers are approximately equal
}
\description{
checks whether two numbers are qpproximately the same to some absolute error
}
