% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutilities.r
\name{assert}
\alias{assert}
\title{assert assertion function to enforce conditions}
\usage{
assert(expr, message = "There was an error!", envir = NULL)
}
\arguments{
\item{expr}{the expression to be evaluated}

\item{message}{character to be displaced in the event of an error

return NULL (invisibly)}
}
\description{
assert assertion function to enforce conditions
}
