% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutilities.r
\name{exit}
\alias{exit}
\title{Function to quit R - to make it the same as Julia/Python}
\usage{
exit(save = "no")
}
\arguments{
\item{save}{character "yes" or "no" passed to q() defaults to "no"}
}
\description{
Function to quit R - to make it the same as Julia/Python
}
