% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutilities.r
\docType{methods}
\name{\%+\%}
\alias{\%+\%}
\alias{\%+\%,character,character-method}
\title{Binary function \%+\% for concetenating characters}
\usage{
x \%+\% y

\S4method{\%+\%}{character,character}(x, y)
}
\arguments{
\item{x}{character on the lhs}

\item{y}{character on the rhs}
}
\description{
Binary function \%+\% for concetenating characters
}
\examples{
         "Hello " \%+\% "World!"


}
