% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_utils.R
\name{is_git_repo_clean}
\alias{is_git_repo_clean}
\title{Check that a git project is clean}
\usage{
is_git_repo_clean(git_repo = ".", stop_if_false = TRUE,
  verbose = FALSE)
}
\description{
Check that a git project has no uncommitted changes or untracked files, and stop if
they are found
}
