% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutilities.r
\name{removeAndUpdate}
\alias{removeAndUpdate}
\title{Function to remove a package and recompile and reinstall it}
\usage{
removeAndUpdate(packageName, path = ".")
}
\description{
Function to remove a package and recompile and reinstall it
}
