% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutilities.r
\name{removeAndUpdatePackage}
\alias{removeAndUpdatePackage}
\title{Function to remove and update a locally available package}
\usage{
removeAndUpdatePackage(package = "saucer")
}
\arguments{
\item{package}{the package name to be passed to remove.packages()
and updatePackage()}
}
\value{
NULL invisibly
}
\description{
Function to remove and update a locally available package
using the updatePackage() function
}
