% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutilities-package.R
\docType{package}
\name{rutilities-package}
\alias{rutilities}
\alias{rutilities-package}
\title{rutilities: Utilities for Phil Dyer}
\description{
Set of functions that I have needed.
    Functions will start out specific
    and be extended if I need to.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PhDyellow/rutilities}
  \item Report bugs at \url{https://github.com/PhDyellow/rutilities/issues}
}

}
\author{
\strong{Maintainer}: Phil Dyer \email{phildyer@protonmail.com}

}
\keyword{internal}
