% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_utils.R
\name{track_system_state}
\alias{track_system_state}
\title{Print state of system}
\usage{
track_system_state()
}
\description{
Outputs a set of details about the system that is currently running
}
\details{
Outputs:
  System time,
  R version
  OS details
}
