% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutilities.r
\name{updatePackage}
\alias{updatePackage}
\title{Function compiles and installs/updates
an R package on the system}
\usage{
updatePackage(package_name, path = ".")
}
\arguments{
\item{package_name}{character for the name of the package - folder name}

\item{path}{character the path where the folder is located}
}
\description{
Function compiles and installs/updates
an R package on the system
}
\examples{
       # updatePackage("saucer")

}
